:fix: review events layer and metrics
