+++ commons/src/main/java/org/eclipse/kapua/commons/localevent/EventHandler.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2021, 2022 Eurotech and/or its affiliates and others
+ * Copyright (c) 2021, 2026 Eurotech and/or its affiliates and others
  *
  * This program and the accompanying materials are made
  * available under the terms of the Eclipse Public License 2.0
@@ -12,43 +12,54 @@
  *******************************************************************************/
 package org.eclipse.kapua.commons.localevent;
 
-import org.eclipse.kapua.commons.metric.CommonsMetric;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import com.codahale.metrics.Counter;
 
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.LinkedBlockingDeque;
 import java.util.concurrent.TimeUnit;
 
 /**
- * @param <O> Who knows?
- * @deprecated since 2.0.0 - this was probably intended to be part of a larger abstraction that never came to fruition - as it is now it just creates confusion.
- * Do not use, will be removed in future releases
+ * Generic class to allow different component to interact with a blocking queue.
+ * This is usable to distribute internal events without involving brokers.
+ * This class provide a consumer that polling for new messages in the queue and a producer method to put new messages in that queue.
+ * O generic object
  */
-@Deprecated
 public abstract class EventHandler<O> {
 
     private static Logger logger = LoggerFactory.getLogger(EventHandler.class);
-    private final CommonsMetric commonsMetric;
 
     private boolean running;
     private ExecutorWrapper executorWrapper;
     private static final int MAX_ONGOING_OPERATION = 10;
-    private static final int POLL_TIMEOUT = 10000;
 
     private BlockingQueue<O> eventQueue = new LinkedBlockingDeque<>(MAX_ONGOING_OPERATION);
     private EventProcessor<O> eventProcessor;
 
-    protected EventHandler(CommonsMetric commonsMetric, String name, long initialDelay, long pollTimeout) {
-        this.commonsMetric = commonsMetric;
+    private Counter enqueuedEvent;
+
+    /**
+     * Default constructor. Warn: doesn't start without start() is called
+     * @param name Event handler name useful to distinguish between multiple instances (and used while logging information)
+     * @param initialDelay initial delay before executor start, after start() is called (in seconds!)
+     * @param pollTimeout poll timeout (in milliseconds!)
+     * @param enqueuedEvent counter for enqueued events
+     * @param dequeuedEvent counter for dequeued events
+     * @param processedEvent counter for processed events
+     */
+    public EventHandler(String name, long initialDelay, long pollTimeout,
+            Counter enqueuedEvent, Counter dequeuedEvent, Counter processedEvent) {
+        this.enqueuedEvent = enqueuedEvent;
         executorWrapper = new ExecutorWrapper(name, () -> {
             while (isRunning()) {
                 try {
-                    O eventBean = eventQueue.poll(POLL_TIMEOUT, TimeUnit.MILLISECONDS);
+                    O eventBean = eventQueue.poll(pollTimeout, TimeUnit.MILLISECONDS);
                     if (eventBean != null) {
-                        commonsMetric.getDequeuedEvent().inc();
+                        dequeuedEvent.inc();
                         eventProcessor.processEvent(eventBean);
-                        commonsMetric.getProcessedEvent().inc();
+                        processedEvent.inc();
                     }
                 } catch (InterruptedException e) {
                     //do nothing...
@@ -64,7 +75,7 @@ public abstract class EventHandler<O> {
 
     public void enqueueEvent(O eventBean) {
         eventQueue.add(eventBean);
-        commonsMetric.getEnqueuedEvent().inc();
+        enqueuedEvent.inc();
     }
 
     public void registerConsumer(EventProcessor<O> eventProcessor) {
